s7_server/requirements.txt
python-snap7==1.3
pyyaml
setuptools